requirements.txt
# SuperVision Project Requirements
# Python 3.10+ recommended

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.3.0

# Data processing
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.6.0
opencv-python>=4.7.0

# Materials science
ase>=3.22.0
pymatgen>=2023.3.0
matminer>=0.9.0

# Deep learning utilities
transformers>=4.30.0
Pillow>=9.4.0

# Utilities
requests>=2.28.0
tqdm>=4.65.0
joblib>=1.2.0
